requirements.txt
langgraph==0.2.28
langchain-openai==0.2.2
gymnasium[box2d,classic_control]==0.29.1
python-dotenv==1.0.1
pydantic==2.9.2
pyyaml==6.0.2
stable-baselines3==2.3.2
torch==2.6.0
tensorboard>=2.9.1
matplotlib
opencv-python
psutil
pillow
rich==14.2.0
